backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.35
asyncmy==0.2.9
pymysql==1.1.1
pymilvus==2.4.7
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
alembic==1.13.2
captcha==0.5.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
markdown==3.7
setuptools>=68.0.0
aiosqlite==0.20.0
